remove demo
