--- wren-ai-service/demo/.env.example
@@ -1,8 +1,0 @@
-bigquery.project-id=
-bigquery.dataset-id=
-bigquery.credentials-key=
-postgres.host=
-postgres.port=
-postgres.database=
-postgres.user=
-postgres.password=
